[Agent] fix class closures after mixin extraction
